Refactor: Apply SOLID principles to split PDF files into modules
