+++ pdf_splitter.spec
@@ -6,7 +6,7 @@ a = Analysis(
     pathex=[],
     binaries=[],
     datas=[],
-    hiddenimports=[],
+    hiddenimports=['pdf_splitter'],
     hookspath=[],
     hooksconfig={},
     runtime_hooks=[],
