+++ pdf_splitter/__init__.py
@@ -1,0 +1,8 @@
+"""
+PDF Splitter - A tool for splitting PDF pages vertically
+"""
+
+from .pdf_processor import PDFProcessor
+from .config import Config
+
+__version__ = '1.0.1'
