+++ version_info.txt
@@ -1,5 +1,6 @@
-VSVersionInfo(
-  ffi=FixedFileInfo(    filevers=(1, 0, 1, 0),
+﻿VSVersionInfo(
+  ffi=FixedFileInfo(
+    filevers=(1, 0, 1, 0),
     prodvers=(1, 0, 1, 0),
     mask=0x3f,
     flags=0x0,
@@ -9,19 +10,16 @@ VSVersionInfo(
     date=(0, 0)
   ),
   kids=[
-    StringFileInfo(
-      [
-        StringTable(
-          u'040904B0',
-          [StringStruct(u'CompanyName', u''),
-           StringStruct(u'FileDescription', u'PDF Page Vertical Splitter'),
-           StringStruct(u'FileVersion', u'1.0.1'),
-           StringStruct(u'InternalName', u'pdf_splitter'),
-           StringStruct(u'LegalCopyright', u'MIT License'),
-           StringStruct(u'OriginalFilename', u'pdf_splitter.exe'),
-           StringStruct(u'ProductName', u'PDF Splitter'),
-           StringStruct(u'ProductVersion', u'1.0.0')])
-      ]),
+    StringFileInfo([StringTable(
+      u'040904b0',
+      [StringStruct(u'CompanyName', u''),
+       StringStruct(u'FileDescription', u'PDF Page Vertical Splitter'),
+       StringStruct(u'FileVersion', u'1.0.1'),
+       StringStruct(u'InternalName', u'pdf_splitter'),
+       StringStruct(u'LegalCopyright', u'MIT License'),
+       StringStruct(u'OriginalFilename', u'pdf_splitter.exe'),
+       StringStruct(u'ProductName', u'PDF Splitter'),
+       StringStruct(u'ProductVersion', u'1.0.1')])]),
     VarFileInfo([VarStruct(u'Translation', [1033, 1200])])
   ]
 )
